backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone-client
langchain
langchain-openai
langchain-community
langchain-pinecone
langchain-text-splitters
langchain-core
beautifulsoup4
pandas
requests